% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{bh}
\alias{bh}
\title{bh}
\usage{
bh(pvals, alpha = 0.05)
}
\arguments{
\item{pvals}{Vector of p-values to be tested.}

\item{alpha}{Number between 0 and 1 indicating level at which to control FDR.}
}
\value{
Indices of p-values to be rejected.
}
\description{
Given a vector of p-values, selects a subset to reject based on the 
Benjamini-Hochberg FDR control procedure.
}
