% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{BWpriorities}
\alias{BWpriorities}
\title{BW Dataset}
\format{
A dataset in the standard DCE format.
}
\source{
not sure
}
\usage{
data(BWpriorities)
}
\description{
Data from ....
}
\references{
not sure
}
\keyword{datasets}
