% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChecks.R
\name{choice_model_vif}
\alias{choice_model_vif}
\title{choice_model_vif}
\usage{
choice_model_vif(processed_data)
}
\arguments{
\item{processed_data}{processed data list}
}
\value{
vif
}
\description{
choice_model_vif
}
