% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMIFileCreator.R
\name{create_emi_file}
\alias{create_emi_file}
\title{create_emi_file}
\usage{
create_emi_file(
  file_name,
  model_type,
  ncovariates = NULL,
  attribute_names = NULL,
  verbose = 1
)
}
\arguments{
\item{file_name}{processed data list}

\item{model_type}{type for base model string}

\item{ncovariates}{need this or name list}

\item{attribute_names}{need this ncovariates}

\item{verbose}{prints more if >0}
}
\value{
creates file
}
\description{
create_emi_file
}
