% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_validation_doc.R
\name{create_validation_doc}
\alias{create_validation_doc}
\title{create_validation_doc}
\usage{
create_validation_doc(validation.data, init.data, preproc.steps = list(),
  data.flags = list(), validation.dir = "")
}
\arguments{
\item{validation.data}{Output of \code{run_DCM} with \code{validation==TRUE}.
A list containing lists, \code{it.pvals}, \code{it.test.stats}, 
\code{it.std.errs}, \code{it.sets}, and \code{it.diff.cor}.}

\item{init.data}{Output of \code{init_DCM}. A list containing a list, 
\code{it.scores}.}

\item{preproc.steps}{Vector containing strings with all preprocessing steps 
performed.}

\item{data.flags}{Vector containing strings with all data flags thrown by 
the algorithm.}

\item{validation.dir}{String giving the directory in which the pdf and txt
docs should be saved.}
}
\value{
None
}
\description{
Creates pdf and txt documents containing validation diagnostics for the DCM 
algorithm.
}
