% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{createConcepts}
\alias{createConcepts}
\title{create concepts.}
\usage{
createConcepts(data_matrix, nmax_choiceset_size = 31)
}
\arguments{
\item{data_matrix}{matrix}

\item{nmax_choiceset_size}{int default 31}
}
\value{
List  of outputs .
}
\description{
create concepts.
}
