% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMI.R
\name{createEMIWorkbook}
\alias{createEMIWorkbook}
\title{Create workbook}
\usage{
createEMIWorkbook(pre_processed_data, model_type, working_folder = NULL)
}
\arguments{
\item{pre_processed_data}{processed data list}

\item{model_type}{string}

\item{working_folder}{string}
}
\value{
name of workbook
}
\description{
Create workbook
}
