% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{DCEpriorities}
\alias{DCEpriorities}
\title{DCE Dataset}
\format{
A dataset in the standard DCE format.
}
\source{
not sure
}
\usage{
data(DCEpriorities)
}
\description{
Data from ....
}
\references{
not sure
}
\keyword{datasets}
