% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCM-package.R
\docType{package}
\name{DCM-package}
\alias{DCM}
\alias{DCM-package}
\title{DCM: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewthomasjones/DCM}
  \item Report bugs at \url{https://github.com/andrewthomasjones/DCM/issues}
}

}
\keyword{internal}
