% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dcm}
\alias{dcm}
\title{Discrete Choice Model (DCM) for Nonignorable Missing Data}
\usage{
dcm(data_frame, regr_formula, regr_fam)
}
\arguments{
\item{data_frame}{A data frame containing the variables in the regression model.}

\item{regr_formula}{An object of class "formula" describing the regression model to be fitted}

\item{regr_fam}{A description of the error distribution and link function to be used in the regression model. 
This can be a character string (e.g. "binomial" or "gaussian") naming a family function, same as the input to glm().}
}
\value{
A "dcm" object containing the following items:
\item{CC}{An object of class "glm" from complete-case regression analysis.}
\item{IPW}{The corresponding coefficient point estimate, bootstrap standard error, confidence interval and p-value from IPW analysis.}
\item{DAT}{A data frame containing the variables in the regression model and the missing data pattern indicator R.}
}
\description{
Implements inverse probability weighting (IPW) with the logit discrete choice nonresponse model (LDCM).
}
\details{
This function implements IPW using LDCM weights as described in Tchetgen Tchetgen et al. (2017). The implementation is based on 
a default linear main effects model of the observed variables in the \emph{r}-th missing data pattern for the log ratio 
of probabilities for observing the \emph{r}-th missing data pattern versus the complete data.
}
\examples{

#The 'airquality' dataset is included in R base and contains nonmonotone missing values
#in the variables Ozone (ppb) and Solar.R (lang). 
out <- dcm(airquality, Temp~Ozone+Solar.R+Wind, gaussian)

}
\references{
Tchetgen Tchetgen, E., Wang, L. and Sun, B. (2017). \href{http://www3.stat.sinica.edu.tw/ss_newpaper/SS-2016-0325_na.pdf}{Discrete Choice Models for Nonmonotone
Nonignorable Missing Data: Identification and Inference}. Statistica Sinica (doi: 10.5705/ss.202016.0325).
}
