% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{estimate_model}
\alias{estimate_model}
\title{estimate things}
\usage{
estimate_model(model_sims, type, precision)
}
\arguments{
\item{model_sims}{simulation to estimate from}

\item{type}{estimation method}

\item{precision}{precision parameter}
}
\value{
results
}
\description{
estimate things
}
