% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{fisher}
\usage{
fisher(r)
}
\arguments{
\item{r}{Number between -1 and 1 to be transformed.}
}
\value{
If \code{-1 < r < 1}, Fisher transformed input. Else, 0.
}
\description{
Performs a Fisher transformation on a number between -1 and 1. If input is 
outside that range, returns 0.
}
