% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_trunc.R
\name{fisher_trunc}
\alias{fisher_trunc}
\title{fisher_trunc}
\usage{
fisher_trunc(r)
}
\arguments{
\item{r}{Number between -1 and 1 to be transformed.}
}
\value{
Fisher transformed input truncated at -1 or 1 if outside the interval
(-1, 1).
}
\description{
Performs a Fisher transformation on a number between -1 and 1, truncating to 
-1 or 1 if the input lies outside of that range.
}
