% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphOutput.R
\name{fixed_model_graph}
\alias{fixed_model_graph}
\title{fixed_model_graph}
\usage{
fixed_model_graph(fitted)
}
\arguments{
\item{fitted}{fitted model fixed only for now}
}
\value{
graph object
}
\description{
fixed_model_graph
}
