% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyDistribution.R
\name{frequencyDistribution_simple}
\alias{frequencyDistribution_simple}
\title{Frequency distribution
this probably shouldnt be exported to be honest}
\usage{
frequencyDistribution_simple(coldd)
}
\arguments{
\item{cs}{data}
}
\value{
fdd
}
\description{
Frequency distribution
this probably shouldnt be exported to be honest
}
