% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{generate_simulation_templates}
\alias{generate_simulation_templates}
\title{creates structure for simulations}
\usage{
generate_simulation_templates(m, p = 2)
}
\arguments{
\item{m}{sample size}

\item{p}{variables}
}
\value{
results
}
\description{
creates structure for simulations
}
