% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModels.R
\name{generateModelMatrices}
\alias{generateModelMatrices}
\title{generateModelMatrices}
\usage{
generateModelMatrices(pre_processed_data, model_type)
}
\arguments{
\item{pre_processed_data}{processed data list}

\item{model_type}{string}
}
\value{
model matrices
}
\description{
generateModelMatrices
}
