% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{gqIntMatrix}
\alias{gqIntMatrix}
\title{Integral matrix}
\usage{
gqIntMatrix(integral_size, nrc)
}
\arguments{
\item{integral_size}{int number steps for integral using Gaussian quadrature}

\item{nrc}{int number of columns}
}
\value{
integral matrix
}
\description{
Integral matrix
}
