% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_DCM.R
\name{init_DCM}
\alias{init_DCM}
\title{init_DCM}
\usage{
init_DCM(M1, M2, k, start = c(), del = c())
}
\arguments{
\item{M1}{p x n1 data matrix for group 1.}

\item{M2}{p x n2 data matrix for group 2.}

\item{k}{Desired size of initial set.}

\item{start}{Desired initial set.}

\item{del}{Vector of variables to be excluded from initialization set.}
}
\value{
List containing,
\itemize{
   \item seed - Vector containing the seed set used.
   \item found - Vector containing the terminating set to be used as initial 
   set for iterative testing procedure.
   \item iterations - Number of iterations performed.
   \item time - Amount of time that the initialization procedure took.
   \item init.sets - List containing set at each iteration.
   \item it.scores - Vector containing the score for the set at each 
   iteration.
   \item init.steps - List containing strings that give information about 
   what steps are performed during initialization.
}
}
\description{
Takes prepared Matrix (row-normalized and optionally Quantile Normalized).
 Looks for high correlation in M1 and low/negative correlation in M2, finds 
 group of size k to be used as initial set in iterative testing procedure.
}
