% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_choicedatasets}
\alias{join_choicedatasets}
\title{join}
\usage{
join_choicedatasets(data1, data2)
}
\arguments{
\item{data1}{processed data list}

\item{data2}{processed data list}
}
\value{
joins two processed objects by cols
}
\description{
join
}
