% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinChoiceDatasets.R
\name{joinChoiceDatasets}
\alias{joinChoiceDatasets}
\title{join}
\usage{
joinChoiceDatasets(data1, data2)
}
\arguments{
\item{data1}{processed data list}

\item{data2}{processed data list}
}
\value{
joins two processed objects by cols. only takes  ID ChoiceSet Choice from data1
variable names must be in format variable_dataset and match across the two datasets
}
\description{
join
}
