% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llCalc.R
\name{llCalc}
\alias{llCalc}
\title{does LL calc
not used - C++ version is}
\usage{
llCalc(working_values, model, processed, draws_matrix)
}
\value{
loglik
}
\description{
does LL calc
not used - C++ version is
}
