% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llCalc.R
\name{llMax}
\alias{llMax}
\title{calculates optimum model fit
not used - C++ version is}
\usage{
llMax(model, processed, ghq_matrix1, nlm_params)
}
\arguments{
\item{model}{model list}

\item{processed}{data list}

\item{ghq_matrix1}{mat}

\item{nlm_params}{list of params}
}
\value{
max loglik
}
\description{
calculates optimum model fit
not used - C++ version is
}
