% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llMax2}
\alias{llMax2}
\title{llMax2
does the maximisation}
\usage{
llMax2(model, processed, draws_matrix)
}
\arguments{
\item{model}{list}

\item{processed}{list}

\item{draws_matrix}{matrix}
}
\value{
opt_results
}
\description{
llMax2
does the maximisation
}
