% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llMax_ghq}
\alias{llMax_ghq}
\title{llMax_ghq
does the maximisation}
\usage{
llMax_ghq(model, processed, ghq_matrix1, nlm_params)
}
\arguments{
\item{model}{list}

\item{processed}{list}

\item{ghq_matrix1}{matrix}

\item{nlm_params}{list of params}
}
\value{
opt_results
}
\description{
llMax_ghq
does the maximisation
}
