% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMI.R
\name{loadEMIWorkbook}
\alias{loadEMIWorkbook}
\title{Loads an existing workbook}
\usage{
loadEMIWorkbook(pre_processed_data, emi_file_name)
}
\arguments{
\item{emi_file_name}{filename string}
}
\value{
model
}
\description{
Loads an existing workbook
}
