% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSim.R
\name{makeSim}
\alias{makeSim}
\title{makeSim}
\usage{
makeSim(p, k, rho, n1, n2, bg = "n", rho2 = 0, varinf = FALSE,
  large_k = 0)
}
\arguments{
\item{p}{Number of variables.}

\item{k}{Size of differentially correlated set.}

\item{rho}{Correlation of set in first group.}

\item{n1}{Number of columns in first data matrix.}

\item{n2}{Number of columns in second data matrix.}

\item{bg}{String indicating what type of background should be used 
(\code{"n"}, \code{"pos"}, or \code{"noisy"}).}

\item{rho2}{Correlation of set in second group.}

\item{varinf}{Boolean, if \code{varinf==TRUE}, sets diagonals of covariance
matrices to random uniform[0,5] values. Else, sets them to one.}

\item{large_k}{Integer describing how size of constant, positive submatrix to
add to covariance matrices.}
}
\value{
List containing
\itemize{
    \item dat1 - Simulated p by n1 data matrix.
    \item dat2 - Simulated p by n2 data matrix.
}
}
\description{
Simulates two data matrices with pre-specified differentially correlated set.
}
