% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVar.R
\name{makeVar}
\alias{makeVar}
\title{makeVar}
\usage{
makeVar(Ui, M_A)
}
\arguments{
\item{Ui}{Normalized row in A for a single group.}

\item{M_A}{Row-normalized matrix containing rows inside of A excluding Uifor 
a single group.}
}
\value{
Number indicating variance of variable Ui.
}
\description{
Estimates variance of the DCM test statistic for a variable inside the set A
using the estimator of Steiger and Hakstian (1982). More details are provided
in section 3.2 of the DCM paper.
}
