% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVars.R
\name{makeVars}
\alias{makeVars}
\title{makeVars}
\usage{
makeVars(Uis, M_A)
}
\arguments{
\item{Uis}{Row-normalized matrix containing rows outside of set A for a 
single group.}

\item{M_A}{Row-normalized matrix containing rows inside of A for a single 
group.}
}
\value{
Vector giving variance estimates for each row outside of set A.
}
\description{
Estimates variance of the DCM test statistic for variables outside the set A
using the estimator of Steiger and Hakstian (1982). More details are provided
in section 3.2 of the DCM paper.
}
