% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModels.R
\name{model_generator}
\alias{model_generator}
\title{generate model}
\usage{
model_generator(pre_processed_data, model_type)
}
\arguments{
\item{pre_processed_data}{processed data list}

\item{model_type}{string}
}
\value{
model
}
\description{
generate model
}
