% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphOutput.R
\name{modelGraph}
\alias{modelGraph}
\title{generateModelMatrices}
\usage{
modelGraph(object)
}
\arguments{
\item{object}{model or results object}

\item{model_type}{string}
}
\value{
model matrices
}
\description{
generateModelMatrices
}
