% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{parameter_labels}
\alias{parameter_labels}
\title{Create parameter labels}
\usage{
parameter_labels(model)
}
\arguments{
\item{model}{model list}
}
\value{
list of parmeter names
}
\description{
Create parameter labels
}
