% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterLabels.R
\name{parameterLabels}
\alias{parameterLabels}
\title{Create parameter labels}
\usage{
parameterLabels(model)
}
\arguments{
\item{model}{model list}
}
\value{
list of parmeter names
}
\description{
Create parameter labels
}
