% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.dcm}
\alias{predict.dcm}
\title{predict.dcm}
\usage{
\method{predict}{dcm}(results, data = NULL)
}
\arguments{
\item{results}{results object}

\item{data}{data to make predictions on}
}
\value{
predictions
}
\description{
predict.dcm
}
