% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData_DCM.R
\name{prepData_DCM}
\alias{prepData_DCM}
\title{prepData_DCM}
\usage{
prepData_DCM(M1, M2, QN = FALSE, resid.full = FALSE, echo = FALSE,
  strict = "low", validation = FALSE)
}
\arguments{
\item{M1}{p x n1 data matrix for group 1.}

\item{M2}{p x n2 data matrix for group 2.}

\item{QN}{Boolean indicating whether quantile normalization should be
performed.}

\item{resid.full}{Boolean indicating whether residualization of the full
matrix should be performed.}

\item{echo}{Boolean indicating whether information should be printed to the
console.}

\item{strict}{String indicating whether rows with low variance should be
filtered from the matrices. See sanitize_DCM.R for usage. (\code{"high"} or
\code{"low"}).}

\item{validation}{Boolean indicating whether validation data should be stored
and returned.}
}
\value{
If \code{validation == TRUE}, list containing,
\itemize{
   \item M1 - Prepared data matrix for group 1.
   \item M2 - Prepared data matrix for group 2.
   \item idcs - Vector of rows which have been included.
   \item preproc.steps - List containing strings that describe preprocessing
   steps that were performed.
   \item data.flags - List containing strings that describe any flags about
   the data that were observed.
}
Else, list containing,
\itemize{
   \item M1 - Prepared data matrix for group 1.
   \item M2 - Prepared data matrix for group 2.
   \item idcs - Vector of rows which have been included.
}
}
\description{
Takes two matrices and prepares them for use in the DCM procedure.
}
