% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{process_sims}
\alias{process_sims}
\title{cleans up data batch of simulations. params need to mathc those used for sim}
\usage{
process_sims(
  big_list,
  data_sets,
  chosen_values,
  precision_levels,
  integral_types,
  models,
  m_list,
  n_sims,
  conf_level = 0.95
)
}
\arguments{
\item{big_list}{list of sim results}

\item{data_sets}{list of dataset types}

\item{chosen_values}{simulation params}

\item{precision_levels}{for estimates}

\item{integral_types}{estimation method}

\item{models}{model types to test}

\item{m_list}{list of sample sizes for sims}

\item{n_sims}{no of sims}

\item{conf_level}{confidence levels for coverage calculations}
}
\value{
list of results
}
\description{
cleans up data batch of simulations. params need to mathc those used for sim
}
