% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantNorm.R
\name{quantNorm}
\alias{quantNorm}
\title{quantNorm}
\usage{
quantNorm(Mat)
}
\arguments{
\item{Mat}{Matrix to be quantile normalized.}
}
\value{
Matrix with rows quantile normalized.
}
\description{
Quantile normalize rows of a given matrix.
}
