% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in.R
\name{read_data}
\alias{read_data}
\title{Read data. This documentation is incomplete.}
\usage{
read_data(filename)
}
\value{
A dataframe of the input data.
}
\description{
Read data. This documentation is incomplete.
}
\examples{
x<-read_data(test.txt)
x<-read_data(test.xlsx)
}
