% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{readData}
\alias{readData}
\title{Read data. This documentation is incomplete.}
\usage{
readData(filename, header = TRUE)
}
\value{
A dataframe of the input data.
}
\description{
Read data. This documentation is incomplete.
}
\examples{
x <- read_data(test.txt)
x <- read_data(test.xlsx)
}
