% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{remove_variable}
\alias{remove_variable}
\title{remove variable This documentation is incomplete.}
\usage{
remove_variable(processed_data, variable)
}
\arguments{
\item{processed_data}{processes data list}

\item{variable}{variable to remove, string}
}
\value{
a new processed data object with everything updated
}
\description{
remove variable This documentation is incomplete.
}
