% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{removeVariables}
\alias{removeVariables}
\title{remove variables This documentation is incomplete.}
\usage{
removeVariables(processed_data, variable, verbose = FALSE)
}
\arguments{
\item{processed_data}{processes data list}

\item{variable}{variable to remove, string or number or vector}

\item{verbose}{TRUE means print more}
}
\value{
a new processed data object with everything updated
}
\description{
remove variables This documentation is incomplete.
}
