% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DCM.R
\name{run_DCM}
\alias{run_DCM}
\title{run_DCM}
\usage{
run_DCM(M1, M2, seed, del = c(), echo = FALSE, alpha = 0.05,
  max.iter = 50, max.size = 1000, validation = FALSE)
}
\arguments{
\item{M1}{p x n1 data matrix for group 1.}

\item{M2}{p x n2 data matrix for group 2.}

\item{seed}{Vector giving initial set of variables.}

\item{del}{Vector containing variables to be ignored.}

\item{echo}{Boolean indicating whether information should be printed to the
console.}

\item{alpha}{Number between 0 and 1 indicating the level at which to control
the false discovery rate at each iteration.}

\item{max.iter}{Maximum number of iterations to be performed.}

\item{max.size}{Maximum size of set to consider.}

\item{validation}{Boolean indicating whether validation data should be stored
and returned.}
}
\value{
If \code{validation==TRUE}, list containing,
\itemize{
   \item found - Vector containing the variables in the differentially 
   correlated set found by the algorithm.
   \item mc1 - Mean correlation of \code{found} in group 1.
   \item mc2 - Mean correlation of \code{found} in group 2.
   \item its - Number of iterations performed.
   \item time - Time elapsed during procedure.
   \item pvals - Vector of p-values produced at last iteration of testing 
   procedure.
   \item startdels - Vector of variables which were ignored.
   \item it.sets - List of vectors containing differentially correlated sets
   at each iteration.
   \item it.pvals - List of vectors containing p-values produced at each
   iteration.
   \item it.test.stats - List of vectors containing test statistics produced
   at each iteration.
   \item it.std.errs - List of vectors containing standard errors produced at
   each iteration.
   \item it.diff.cor - Vector of mean differential correlations at each 
   iteration.
   \item reason.for.terminating - String describing reason why the algorithm
   terminated.
   \item flags - Vector of strings containing any warnings that arose during
   the search.
}
Else, list containing,
\itemize{
   \item found - Vector containing the variables in the differentially 
   correlated set found by the algorithm.
   \item mc1 - Mean correlation of \code{found} in group 1.
   \item mc2 - Mean correlation of \code{found} in group 2.
   \item its - Number of iterations performed.
   \item time - Time elapsed during procedure.
   \item pvals - Vector of p-values produced at last iteration of testing 
   procedure.
   \item startdels - Vector of variables which were ignored.
   \item reason.for.terminating - String describing reason why the algorithm
   terminated.
}
}
\description{
Searches for a differentially correlated set of variables in matrices M1 and
M2 using an iterative testing procedure.
}
