% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llCalc_TMB.R
\name{run_model_TMB}
\alias{run_model_TMB}
\title{Runs Model}
\usage{
run_model_TMB(model, verbose = FALSE)
}
\arguments{
\item{model}{model list}

\item{verbose}{default FALSE}
}
\value{
fitted model.
}
\description{
Runs Model
}
