% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{run_sims}
\alias{run_sims}
\title{runs a batch of simulations}
\usage{
run_sims(
  data_sets,
  chosen_values,
  precision_levels,
  integral_types,
  models,
  m_list,
  n_sims,
  p = 2,
  file = NULL
)
}
\arguments{
\item{data_sets}{list of dataset types}

\item{chosen_values}{simulation params}

\item{precision_levels}{for estimates}

\item{integral_types}{estimation method}

\item{models}{model types to test}

\item{m_list}{list of sample sizes for sims}

\item{n_sims}{no of sims}

\item{p}{number of columns}

\item{file}{outfile for save}
}
\value{
list of results
}
\description{
runs a batch of simulations
}
