% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_DCM.R
\name{sanitize_DCM}
\alias{sanitize_DCM}
\title{sanitize_DCM}
\usage{
sanitize_DCM(M1, M2, strict = "high", validation = FALSE)
}
\arguments{
\item{M1}{p x n1 data matrix for group 1.}

\item{M2}{p x n2 data matrix for group 2.}

\item{strict}{String indicating whether low variance rows should be removed
(\code{"high"} or \code{"low"})}

\item{validation}{Boolean indicating whether validation data should be 
stored and returned.}
}
\value{
If \code{validation == TRUE}, list containing,
\itemize{
   \item killrows - Vectors of rows to be removed from M1 and M2.
   \item na1 - Vector of rows in M1 with na's.
   \item na2 - Vector of rows in M2 with na's.
   \item lb1 - Vector of rows in M1 with more than \code{ncol(M1)/10} zeroes.
   \item lb2 - Vector of rows in M2 with more than \code{ncol(M2)/10} zeroes.
   \item ub1 - Vector of rows in M1 with more than \code{ncol(M1)/10} ones.
   \item ub1 - Vector of rows in M2 with more than \code{ncol(M2)/10} ones.
   \item var1 - Vector of rows in M1 with variance less than 
   \code{1/(100*ncol(M1))}.
   \item var2 - Vector of rows in M2 with variance less than
   \code{1/(100*ncol(M2))}.
}
Else, vector of rows to be removed.
}
\description{
Takes data matrices and returns rows with na's, too many zeroes, too many 
ones, and low variances.
}
