% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{scores}
\alias{scores}
\title{does sims
not used - C++ version is}
\usage{
scores(chosen_values, model, processed)
}
\arguments{
\item{chosen_values}{vector}

\item{model}{model to base simulations on}

\item{processed}{data to base simulations on}
}
\value{
scores
}
\description{
does sims
not used - C++ version is
}
