% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{simualte data}
\usage{
simulate_dataset(processed, model_type, chosen_values, easy_guess = FALSE)
}
\arguments{
\item{processed}{data to base simulations on}

\item{model_type}{model_type}

\item{chosen_values}{sim params}

\item{easy_guess}{guess true values}
}
\value{
sims
}
\description{
simualte data
}
