% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lv.R
\name{simulate_lv}
\alias{simulate_lv}
\title{Simulate latent variable}
\usage{
simulate_lv(X, betas, taus, zetas, n)
}
\arguments{
\item{n}{}
}
\description{
Simulate latent variable
}
\examples{
n <- 10e3
X <- data.frame(x1 = rnorm(n), x2 = rnorm(n))
betas <- c(beta1 = 1, beta2 = 2)
taus <- c(tau1 = -1, tau2 = 1)
zetas <- c(zeta1 = 1, zeta2 = 2)
simulate_lv(X, betas, taus, zetas, n)
}
