% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ol.R
\name{simulate_ol}
\alias{simulate_ol}
\title{Simulate ordered logit data}
\usage{
simulate_ol(X, betas, taus, n)
}
\arguments{
\item{X}{data.frame (or matrix)}

\item{betas}{named coef vector}

\item{taus}{named thresholds}

\item{n}{number of obs}
}
\description{
Can be combined with \code{simulate_lv()} to simulate hybrid choice data
}
\examples{
n <- 10e3
X <- data.frame(x1 = rnorm(n), x2 = rnorm(n))
betas <- c(beta1 = 1, beta2 = 2)
taus <- c(tau1 = -1, tau2 = 1) # will generate 3 choices
simulate_ol(X, betas, taus, n)
}
