% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{simulation}
\alias{simulation}
\title{does sims}
\usage{
simulation(chosen_values, model, processed)
}
\arguments{
\item{chosen_values}{vector}

\item{model}{model to base simulations on}

\item{processed}{data to base simulations on}
}
\value{
loglik
}
\description{
does sims
}
