% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize.R
\name{stdize}
\alias{stdize}
\title{stdize}
\usage{
stdize(Mat)
}
\arguments{
\item{mat}{Data matrix to be row-standardized.}
}
\value{
Row-standardized matrix.
}
\description{
Standardize matrix to have mean=0 and sum of squares=1, by rows.
}
