% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtilities.R
\name{summariseModelList}
\alias{summariseModelList}
\title{Summary list}
\usage{
summariseModelList(list_of_models)
}
\arguments{
\item{list_of_models}{list of model list}
}
\value{
ltable of summary info
}
\description{
Summary list
}
