% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{cvError}
\alias{cvError}
\title{cvError}
\usage{
cvError(raw_dataset, k = 0, type = "fixed")
}
\arguments{
\item{raw_dataset}{raw_dataset}

\item{k}{cv folds, 0 is LOOCV}
}
\value{
CV CE Loss
}
\description{
cvError
}
