% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCM.R
\name{DCM}
\alias{DCM}
\title{Data Converter Module}
\usage{
DCM()
}
\value{
DCM is a graphic interface
}
\description{
Data Converter Module (DCM) converts the dataset format from split into stack and to the reverse.

You can learn more about this package at:
http://www.uv.mx/personal/nehuerta/dcm/
}
\details{
The module has two functionalities STACK and SPLIT.

STACK

In this mode the package receives a matrix where the rows are typically the judges or customers that evaluate a product. The first column is the product code, the second column is the ID of the judge and the rest of the columns contain the data of the sensory attributes. The program allows replicated data.

The output of the program is a matrix of stacked data; the first column gives the product code, the second column the ID of the judge, the third column the name of the attribute and the fourth column the value provided by the judge. This matrix is given in a csv file.

SPLIT

The input for this functionality is a matrix that has four columns: Product code, ID of a judge, attribute name and data.

The output is a matrix where the rows correspond to product code-judge combination and the columns provide the sensory characteristics ordered alphabetically or not. This matrix is produced in a csv file.

In both cases stack and split, empty spaces are kept that way.
}
\examples{
\dontrun{
##Install package
library(DCM)
##Call the package
DCM()
}

}
\author{
Delphi Intelligence
}

