% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCM.R
\name{DCM}
\alias{DCM}
\title{DCM}
\usage{
DCM(M1, M2, max.groups = 5, max.iter = 50, max.time = 1,
  est.size = 100, max.size = 1000, alpha = 0.05, start = c(),
  QN = FALSE, resid.full = FALSE, echo = FALSE, strict = "low",
  validation = FALSE, validation.dir = "")
}
\arguments{
\item{M1}{p x n1 data matrix for first group.}

\item{M2}{p x n2 data matrix for second group.}

\item{max.groups}{Maximum number of DC sets to search for.}

\item{max.iter}{Maximum number of iterations to perform at each search.}

\item{max.time}{Maximum number of hours to run algorithm for.}

\item{est.size}{Integer giving size of initial set.}

\item{max.size}{Integer giving maximum set size to consider.}

\item{alpha}{Number between 0 and 1 giving level at which to control false
discovery rate at each iteration.}

\item{start}{Vector of variables to use as initial set.}

\item{QN}{Boolean indicating whether data matrices should be quantile 
normalized.}

\item{resid.full}{Boolean indicating whether full data matrices should be
residualized. Set to true if worried about systematic correlation differences
between the two groups.}

\item{echo}{Boolean indicating whether information should be printed during
runtime.}

\item{strict}{String indicating whether variables with low variances should 
be removed (\code{"high"} or \code{"low"}).}

\item{validation}{Boolean indicating whether validation data should be 
collected and compiled into pdf and txt documents.}

\item{validation.dir}{Directory in which validation documents should be 
saved.}
}
\value{
If \code{validation==TRUE}, list containing
\itemize{
   \item DC_sets - List of vectors containing differentially correlated sets
   found from each search.
   \item iterations - List containing number of iterations from each search.
   \item time - List containing duration in seconds of each search.
   \item meanCor1 - List containing mean correlation in group 1 of the found
   set.
   \item meanCor2 - List containing mean correlation in group 2 of the found
   set.
   \item it.pvals - List of lists containing p-values from each iteration of 
   each search.
   \item it.test.stats - List of lists containing test statistics from each
   iteration of each search.
   \item it.std.errs - List of lists containing standard errors from each 
   iteration of each search.
   \item it.sets - List of lists containing sets from each iteration of each
   search.
   \item inits - List of lists containing initialization data from each 
   search.
   \item indices - Vector of indices indicating rows of the data that were 
   used.
}
Else, list containing
\itemize{
   \item DC_sets - List of vectors containing differentially correlated sets
   found from each search.
   \item iterations - List containing number of iterations from each search.
   \item time - List containing duration in seconds of each search.
   \item meanCor1 - List containing mean correlation in group 1 of the found
   set.
   \item meanCor2 - List containing mean correlation in group 2 of the found
   set.
   \item indices - Vector of indices indicating rows of the data that were 
   used.
}
}
\description{
Given two data matrices, runs the DCM algorithm, searching for sets of
variables with high correlation in the first group and no/less correlation in 
the second group.
}
