% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{drawsMatrix}
\alias{drawsMatrix}
\title{Draws matrix}
\usage{
drawsMatrix(ndraws, nrc, shuffle = TRUE)
}
\arguments{
\item{ndraws}{int number of draws}

\item{nrc}{int number of columns}

\item{shuffle}{deafult TRUE}
}
\value{
draws matrix
}
\description{
Draws matrix
}
