% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{drawsMatrix}
\alias{drawsMatrix}
\title{Draws matrix}
\usage{
drawsMatrix(ndraws, nrc)
}
\arguments{
\item{ndraws}{int number of draws}

\item{nrc}{int number of columns}
}
\value{
draws matrix
}
\description{
Draws matrix
}
