% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyDistribution.R
\name{frequencyDistribution}
\alias{frequencyDistribution}
\title{Frequency distribution
this probably shouldnt be exported to be honest this whole file needs refactoring}
\usage{
frequencyDistribution(cs)
}
\arguments{
\item{cs}{data}
}
\value{
fdd
}
\description{
Frequency distribution
this probably shouldnt be exported to be honest this whole file needs refactoring
}
