% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModels.R
\name{generate_model_matrices}
\alias{generate_model_matrices}
\title{generate_model_matrices}
\usage{
generate_model_matrices(pre_processed_data, model_type)
}
\arguments{
\item{pre_processed_data}{processed data list}

\item{model_type}{string}
}
\value{
model matrices
}
\description{
generate_model_matrices
}
