% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llCalc.R
\name{llCalc}
\alias{llCalc}
\title{does LL calc
not used - C++ version is}
\usage{
llCalc(working_values, model, processed, ghq_matrix1)
}
\arguments{
\item{working_values}{vector}

\item{model}{model list}

\item{processed}{data list}

\item{ghq_matrix1}{mat}
}
\value{
loglik
}
\description{
does LL calc
not used - C++ version is
}
