% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llCalc3}
\alias{llCalc3}
\title{llCalc3
does the calc
actually llCalc3a does, this is same but other cant export}
\usage{
llCalc3(working_values, model, processed, gq_int_matrix)
}
\arguments{
\item{working_values}{vector}

\item{model}{List}

\item{processed}{List}

\item{gq_int_matrix}{matrix}
}
\value{
loglike
}
\description{
llCalc3
does the calc
actually llCalc3a does, this is same but other cant export
}
