% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llCalc.R
\name{llMax}
\alias{llMax}
\title{calculates optimum model fit
not used - C++ version is}
\usage{
llMax(model, processed, draws_matrix)
}
\value{
max loglik
}
\description{
calculates optimum model fit
not used - C++ version is
}
