% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llMax2}
\alias{llMax2}
\title{llMax2
does the maximisation}
\usage{
llMax2(model, processed, gq_int_matrix, nlm_params)
}
\arguments{
\item{model}{list}

\item{processed}{list}

\item{gq_int_matrix}{matrix}

\item{nlm_params}{list of params}
}
\value{
opt_results
}
\description{
llMax2
does the maximisation
}
