% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMI.R
\name{loadEMIWorkbook}
\alias{loadEMIWorkbook}
\title{Loads an existing workbook}
\usage{
loadEMIWorkbook(pre_processed_data, emi_file_name, verbose = FALSE)
}
\arguments{
\item{pre_processed_data}{processed data list}

\item{emi_file_name}{filename string}

\item{verbose}{default FALSE}
}
\value{
model
}
\description{
Loads an existing workbook
}
