% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Runs Model}
\usage{
runModel(model, model_name = "name", ndraws = 1000)
}
\arguments{
\item{model}{model list}

\item{model_name}{string default "name"}

\item{ndraws}{default 1000}
}
\value{
fitted model.
}
\description{
Runs Model
}
