% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Runs Model}
\usage{
runModel(
  model,
  model_name = "name",
  verbose = 0,
  gradtol = 1e-06,
  stepmax = NULL,
  steptol = 1e-06,
  ghq_steps = 1000,
  dev_mode = "orig",
  ghq_size = 3,
  shuffle = TRUE
)
}
\arguments{
\item{model}{model list}

\item{model_name}{string default "name"}

\item{verbose}{default 0}

\item{gradtol}{default 1e-6}

\item{stepmax}{default as in nlm}

\item{steptol}{default 1e-6}

\item{ghq_steps}{default 1000, formerly known asn draws}

\item{dev_mode}{default "orig"}

\item{ghq_size}{4}

\item{shuffle}{default TRUE, for testing purposes}
}
\value{
fitted model.
}
\description{
Runs Model
}
