% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Runs Model}
\usage{
runModel(
  model,
  verbose = FALSE,
  gradtol = 1e-06,
  stepmax = NULL,
  steptol = 1e-06,
  integral_type = NULL,
  ghq_size = NULL,
  draws = NULL
)
}
\arguments{
\item{model}{model list}

\item{verbose}{default FALSE}

\item{gradtol}{default 1e-6}

\item{stepmax}{default as in nlm}

\item{steptol}{default 1e-6}

\item{integral_type}{NULL}

\item{ghq_size}{NUL}

\item{draws}{NULL}
}
\value{
fitted model.
}
\description{
Runs Model
}
