% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{select_variables}
\alias{select_variables}
\title{select variables This documentation is incomplete.}
\usage{
select_variables(processed_data, variable, verbose = 0)
}
\arguments{
\item{processed_data}{processes data list}

\item{variable}{variable to remove, string or number or vector}

\item{verbose}{1 means print more, 0 no print, default 1}
}
\value{
a new processed data object with everything updated
}
\description{
select variables This documentation is incomplete.
}
