% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{setUp}
\alias{setUp}
\title{Read data. This documentation is incomplete.}
\usage{
setUp(filename, header = TRUE)
}
\value{
A list of the processed data.
}
\description{
Read data. This documentation is incomplete.
}
\examples{
x<-setUp(test.txt, header=TRUE)
x<-setUp(test.xlsx)
}
