% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{setUp}
\alias{setUp}
\title{Read data. This documentation is incomplete.}
\usage{
setUp(data, header = TRUE)
}
\value{
A list of the processed data.
}
\description{
Read data. This documentation is incomplete.
}
